perf: switch rarity simulation to segmented memory
This introduces the engine::memory::VirtualMemory structure which
splits the linear memory into segments of fixed size. This will
drastically reduce overall memory consumption, state creation time, and
scoring time. Note that the engine execution time can increase because
allocation is now amortized as part of engine execution.

As part of this change we also switch to memory sizes specified at the
command line being interpreted as MiB instead of MB. This also ensures
memory sizes are multiples of the segment size.