fix: profiler no longer miscounts solver times.
This moves the allocation time of a new profiler when forking states to
an earlier position (into the `with_snapshot` method). This prevents the
solver time for deciding whether a branch is reachable or not from being
lost. Furthermore, we also need to allocate a new profiler when a branch
is skipped, to avoid falsely double-counting the stale profiler.