refactor: simplify `Qubit` struct handling
This simplifies the `Qubit` struct used to identify a qubit variable
when generating a Qubo. Since the qubit is identified directly via its
unique name (i.e. a `u64` value) and we do not form graph structures
with qubits, there is no need to count references to them or to manage
their memeory. We simply pass `Qubit` by value now.

This is a pure refactoring, no functional changes intended.