feat: only print model to stdout if requested
This changes the default invocation of beator to no longer print models
to `stdout` un less requested to do so. To print models to `stdout` one
has to explicitly pass `--output=''` or `--output=-` now. This change in
turn allows us to skip renumbering models if they are not required to be
printed. It is useful when all we want to do is measure time spent on
interactive solver queries.