Fix Emulator: div traced as "divu"
"exec_divu" calls "trace_rtype("divu"...).
"exec_div" also calls "trace_rtype("divu"...) but I believe it should only pass "div" as param.

I came upon this while further studying the code. I am not confident enough yet to 100% call this a typo, since maybe it is intended. But I have not seen this behaviour with other instructions so I would assume it is, in fact, a typo (assuming this is done for logging purposes).