refactor: allow reuse of `CNFBuilder` for SAT solvers
This parametrizes the existing `CNFBuilder` such that it can be reused
for other interfaces. It will allow us to build CNF representations of
arbitrary fragments of a `GateModel` directly in the individual format
of various SAT solvers. We can hence translate directly into the API of
each SAT solver without going through an intermediate representation.